+++ utilities/obfuscation.ts
@@ -1,7 +1,7 @@
-import { Connection } from "./connection/connection.ts";
-import { concat } from "./utilities/buffer.ts";
-import { readBufferFromBigInt } from "./utilities/tl.ts";
-import { ctr256 } from "./deps.ts";
+import { Connection } from "../connection/connection.ts";
+import { concat } from "../utilities/buffer.ts";
+import { readBufferFromBigInt } from "../utilities/tl.ts";
+import { ctr256 } from "../deps.ts";
 
 export async function getObfuscationParameters(
   protocol: number,
